Add gstat wrapper script
Running nix develop produces a valid environment, but trying to build a
firmware results in this error:

  ** (Mix) gstat (coreutils) is required by the Nerves tooling.

Nerves uses a preflight check on MacOS and looks for gstat.

The problem is that on Nix the coreutils package doesn't prefix utils
with g as there's no risk of clashing with system built-ins.

This commit works around the issue by creating a gstat wrapper script
and adding it to the PATH, so that everything can work as expected.